Add files via upload
The Simon game  that challenges players to repeat a sequence of lights and sounds: 
How to play
Players press colored buttons on the game unit to repeat the sequence of lights and sounds that Simon creates. The game starts off with a simple sequence and gets more complex as players progress. Players earn points for each successful repetition, and lose the game if they make a mistake or time runs out.